refactor variable handling and builtins
