+++ src/builtins.h
@@ -38,16 +38,7 @@ Command *get_function(const char *name);
 void remove_function(const char *name);
 void load_functions(void);
 void free_functions(void);
-const char *get_shell_var(const char *name);
-char **get_shell_array(const char *name, int *len);
-void set_shell_var(const char *name, const char *value);
-void set_shell_array(const char *name, char **values, int count);
-void unset_shell_var(const char *name);
-void free_shell_vars(void);
-void push_local_scope(void);
-void pop_local_scope(void);
 int builtin_local(char **args);
-
 int builtin_break(char **args);
 int builtin_continue(char **args);
 int builtin_cond(char **args);
