+++ src/builtins_vars.c
@@ -15,6 +15,7 @@
 #include "func_exec.h"
 #include "scriptargs.h"
 #include "arith.h"
+#include "vars.h"
 #include "options.h"
 #include <stdio.h>
 #include <stdlib.h>
@@ -25,290 +26,6 @@
 
 extern int last_status;
 
-struct var_entry {
-    char *name;
-    char *value;        /* scalar value or NULL when array is used */
-    char **array;       /* NULL for scalar variables */
-    int array_len;
-    struct var_entry *next;
-};
-
-static struct var_entry *shell_vars = NULL;
-
-struct readonly_entry {
-    char *name;
-    struct readonly_entry *next;
-};
-
-static struct readonly_entry *readonly_vars = NULL;
-
-static int is_readonly(const char *name)
-{
-    for (struct readonly_entry *r = readonly_vars; r; r = r->next) {
-        if (strcmp(r->name, name) == 0)
-            return 1;
-    }
-    return 0;
-}
-
-static void add_readonly(const char *name)
-{
-    if (is_readonly(name))
-        return;
-    struct readonly_entry *r = malloc(sizeof(*r));
-    if (!r)
-        return;
-    r->name = strdup(name);
-    if (!r->name) {
-        free(r);
-        return;
-    }
-    r->next = readonly_vars;
-    readonly_vars = r;
-}
-
-struct local_var {
-    char *name;
-    char *value;
-    char **array;
-    int array_len;
-    int had_shell;
-    int had_env;
-    char *env_val;
-    struct local_var *next;
-};
-
-struct local_frame {
-    struct local_var *vars;
-    struct local_frame *next;
-};
-
-static struct local_frame *local_stack = NULL;
-
-static struct local_var *find_local_var(struct local_frame *f, const char *name) {
-    for (struct local_var *v = f ? f->vars : NULL; v; v = v->next) {
-        if (strcmp(v->name, name) == 0)
-            return v;
-    }
-    return NULL;
-}
-
-void push_local_scope(void) {
-    struct local_frame *f = calloc(1, sizeof(*f));
-    if (!f)
-        return;
-    f->next = local_stack;
-    local_stack = f;
-}
-
-void pop_local_scope(void) {
-    if (!local_stack)
-        return;
-    struct local_frame *f = local_stack;
-    local_stack = f->next;
-    while (f->vars) {
-        struct local_var *v = f->vars;
-        f->vars = v->next;
-        if (v->had_shell) {
-            if (v->array)
-                set_shell_array(v->name, v->array, v->array_len);
-            else
-                set_shell_var(v->name, v->value ? v->value : "");
-        } else {
-            unset_shell_var(v->name);
-        }
-        if (v->had_env)
-            setenv(v->name, v->env_val ? v->env_val : "", 1);
-        else
-            unsetenv(v->name);
-        free(v->name);
-        free(v->value);
-        if (v->array) {
-            for (int i = 0; i < v->array_len; i++)
-                free(v->array[i]);
-            free(v->array);
-        }
-        free(v->env_val);
-        free(v);
-    }
-    free(f);
-}
-
-/*
- * Look up a scalar or array variable by name.
- * Returns a pointer to the stored value.  For arrays the first element is
- * returned.  The returned string belongs to the shell and must not be freed.
- * NULL is returned when the variable does not exist.
- */
-const char *get_shell_var(const char *name) {
-    for (struct var_entry *v = shell_vars; v; v = v->next) {
-        if (strcmp(v->name, name) == 0) {
-            if (v->value)
-                return v->value;
-            if (v->array && v->array_len > 0)
-                return v->array[0];
-            return NULL;
-        }
-    }
-    return NULL;
-}
-
-/*
- * Retrieve the array stored under the given name.  The optional len argument
- * is filled with the number of elements.  NULL is returned if no such array
- * exists.
- */
-char **get_shell_array(const char *name, int *len) {
-    for (struct var_entry *v = shell_vars; v; v = v->next) {
-        if (strcmp(v->name, name) == 0 && v->array) {
-            if (len) *len = v->array_len;
-            return v->array;
-        }
-    }
-    if (len) *len = 0;
-    return NULL;
-}
-
-/*
- * Store a scalar variable.  Existing values or arrays with the same name are
- * freed and replaced.  Allocation failures result in a printed error and the
- * variable is left untouched.
- */
-void set_shell_var(const char *name, const char *value) {
-    if (is_readonly(name)) {
-        fprintf(stderr, "%s: readonly variable\n", name);
-        return;
-    }
-    for (struct var_entry *v = shell_vars; v; v = v->next) {
-        if (strcmp(v->name, name) == 0) {
-            if (v->array) {
-                for (int i = 0; i < v->array_len; i++)
-                    free(v->array[i]);
-                free(v->array);
-                v->array = NULL;
-                v->array_len = 0;
-            }
-            free(v->value);
-            v->value = strdup(value);
-            if (opt_allexport)
-                setenv(name, v->value ? v->value : "", 1);
-            return;
-        }
-    }
-    struct var_entry *v = malloc(sizeof(struct var_entry));
-    if (!v) { perror("malloc"); return; }
-    v->name = strdup(name);
-    v->value = strdup(value);
-    v->array = NULL;
-    v->array_len = 0;
-    v->next = shell_vars;
-    shell_vars = v;
-    if (opt_allexport)
-        setenv(name, value ? value : "", 1);
-}
-
-/*
- * Replace or create an array variable.  Existing scalar or array data is
- * freed before the new values are duplicated.  The count parameter indicates
- * the number of elements in 'values'.
- */
-void set_shell_array(const char *name, char **values, int count) {
-    if (is_readonly(name)) {
-        fprintf(stderr, "%s: readonly variable\n", name);
-        return;
-    }
-    for (struct var_entry *v = shell_vars; v; v = v->next) {
-        if (strcmp(v->name, name) == 0) {
-            if (v->value) {
-                free(v->value);
-                v->value = NULL;
-            }
-            if (v->array) {
-                for (int i = 0; i < v->array_len; i++)
-                    free(v->array[i]);
-                free(v->array);
-            }
-            v->array = calloc(count, sizeof(char *));
-            if (v->array) {
-                for (int i = 0; i < count; i++)
-                    v->array[i] = strdup(values[i]);
-            }
-            v->array_len = count;
-            return;
-        }
-    }
-    struct var_entry *v = calloc(1, sizeof(struct var_entry));
-    if (!v) { perror("malloc"); return; }
-    v->name = strdup(name);
-    v->value = NULL;
-    v->array = calloc(count, sizeof(char *));
-    if (v->array) {
-        for (int i = 0; i < count; i++)
-            v->array[i] = strdup(values[i]);
-    }
-    v->array_len = count;
-    v->next = shell_vars;
-    shell_vars = v;
-}
-
-/*
- * Remove a variable or array from the shell table.  All memory associated
- * with the variable is freed.  If the name does not exist nothing happens.
- */
-void unset_shell_var(const char *name) {
-    if (is_readonly(name)) {
-        fprintf(stderr, "%s: readonly variable\n", name);
-        return;
-    }
-    struct var_entry *prev = NULL;
-    for (struct var_entry *v = shell_vars; v; prev = v, v = v->next) {
-        if (strcmp(v->name, name) == 0) {
-            if (prev)
-                prev->next = v->next;
-            else
-                shell_vars = v->next;
-            free(v->name);
-            free(v->value);
-            if (v->array) {
-                for (int i = 0; i < v->array_len; i++)
-                    free(v->array[i]);
-                free(v->array);
-            }
-            free(v);
-            return;
-        }
-    }
-}
-
-/*
- * Free all variables stored by the shell.  This is used when exiting or
- * reinitialising the interpreter.
- */
-void free_shell_vars(void) {
-    struct var_entry *v = shell_vars;
-    while (v) {
-        struct var_entry *n = v->next;
-        free(v->name);
-        free(v->value);
-        if (v->array) {
-            for (int i = 0; i < v->array_len; i++)
-                free(v->array[i]);
-            free(v->array);
-        }
-        free(v);
-        v = n;
-    }
-    shell_vars = NULL;
-
-    struct readonly_entry *r = readonly_vars;
-    while (r) {
-        struct readonly_entry *n = r->next;
-        free(r->name);
-        free(r);
-        r = n;
-    }
-    readonly_vars = NULL;
-}
 
 /*
  * Implements the `shift` builtin which discards the first n positional
@@ -439,144 +156,6 @@ int builtin_set(char **args) {
     return 1;
 }
 
-/* ---- helper functions for builtin_read -------------------------------- */
-
-/*
- * Parse options for the `read` builtin.  On success `idx` is set to the first
- * variable argument and 0 is returned.  A return value of -1 indicates a
- * usage error in the argument list.
- */
-static int parse_read_options(char **args, int *raw, const char **array_name,
-                              int *idx) {
-    int i = 1;
-    *raw = 0;
-    *array_name = NULL;
-
-    if (args[i] && strcmp(args[i], "-r") == 0) {
-        *raw = 1;
-        i++;
-    }
-
-    if (args[i] && strcmp(args[i], "-a") == 0 && args[i + 1]) {
-        *array_name = args[i + 1];
-        i += 2;
-    }
-
-    if (!args[i])
-        return -1;
-
-    *idx = i;
-    return 0;
-}
-
-/*
- * Split a line read by the `read` builtin into words for array assignment.
- * The returned array is heap allocated and must be freed by the caller.  The
- * number of elements is stored in *count.  NULL is returned on allocation
- * failure.
- */
-static char **split_array_values(char *line, int *count) {
-    char **vals = NULL;
-    *count = 0;
-    char *p = line;
-    while (*p) {
-        while (*p == ' ' || *p == '\t')
-            p++;
-        if (*p == '\0')
-            break;
-        char *start = p;
-        while (*p && *p != ' ' && *p != '\t')
-            p++;
-        if (*p)
-            *p++ = '\0';
-        vals = realloc(vals, sizeof(char *) * (*count + 1));
-        if (!vals)
-            return NULL;
-        vals[*count] = strdup(start);
-        (*count)++;
-    }
-    return vals;
-}
-
-/*
- * Helper used by `read` in variable mode.  Splits the line into words and
- * assigns them to the provided variable names starting at index `idx`.
- */
-static void assign_read_vars(char **args, int idx, char *line) {
-    int var_count = 0;
-    for (int i = idx; args[i]; i++)
-        var_count++;
-
-    char *p = line;
-    for (int i = 0; i < var_count; i++) {
-        while (*p == ' ' || *p == '\t')
-            p++;
-        char *val = p;
-        if (i < var_count - 1) {
-            while (*p && *p != ' ' && *p != '\t')
-                p++;
-            if (*p)
-                *p++ = '\0';
-        }
-        set_shell_var(args[idx + i], val);
-    }
-}
-
-/*
- * Implementation of the `read` builtin.  Reads a line from standard input and
- * assigns it to shell variables or an array.  The return status is placed in
- * `last_status` and the function itself always returns 1.
- */
-int builtin_read(char **args) {
-    int raw = 0;
-    const char *array_name = NULL;
-    int idx;
-    if (parse_read_options(args, &raw, &array_name, &idx) != 0) {
-        fprintf(stderr, "usage: read [-r] [-a NAME] NAME...\n");
-        last_status = 1;
-        return 1;
-    }
-
-    char line[MAX_LINE];
-    if (!fgets(line, sizeof(line), stdin)) {
-        last_status = 1;
-        return 1;
-    }
-
-    size_t len = strlen(line);
-    if (len && line[len - 1] == '\n')
-        line[--len] = '\0';
-
-    if (!raw) {
-        char *src = line;
-        char *dst = line;
-        while (*src) {
-            if (*src == '\\' && src[1]) {
-                src++;
-                *dst++ = *src++;
-            } else {
-                *dst++ = *src++;
-            }
-        }
-        *dst = '\0';
-    }
-
-    int array_mode = array_name != NULL;
-    if (array_mode) {
-        int count = 0;
-        char **vals = split_array_values(line, &count);
-        if (!vals)
-            count = 0;
-        set_shell_array(array_name, vals, count);
-        for (int i = 0; i < count; i++)
-            free(vals[i]);
-        free(vals);
-    } else {
-        assign_read_vars(args, idx, line);
-    }
-    last_status = 0;
-    return 1;
-}
 
 /*
  * Evaluate arithmetic expressions supplied to the `let` builtin.  The result
@@ -689,42 +268,6 @@ int builtin_readonly(char **args) {
     return 1;
 }
 
-static void record_local_var(const char *name) {
-    if (!local_stack)
-        return;
-    if (find_local_var(local_stack, name))
-        return;
-    struct local_var *lv = calloc(1, sizeof(*lv));
-    if (!lv)
-        return;
-    lv->name = strdup(name);
-    const char *val = get_shell_var(name);
-    int len = 0;
-    char **arr = get_shell_array(name, &len);
-    if (arr) {
-        lv->array = calloc(len, sizeof(char *));
-        if (lv->array)
-            for (int i = 0; i < len; i++)
-                lv->array[i] = strdup(arr[i]);
-        lv->array_len = len;
-        lv->had_shell = 1;
-    } else if (val) {
-        lv->value = strdup(val);
-        lv->had_shell = 1;
-    } else {
-        lv->had_shell = 0;
-    }
-    const char *e = getenv(name);
-    if (e) {
-        lv->env_val = strdup(e);
-        lv->had_env = 1;
-    } else {
-        lv->had_env = 0;
-    }
-    lv->next = local_stack->vars;
-    local_stack->vars = lv;
-}
-
 static char **parse_array_values(const char *val, int *count) {
     *count = 0;
     char *body = strndup(val + 1, strlen(val) - 2);
@@ -786,184 +329,5 @@ int builtin_local(char **args) {
     return 1;
 }
 
-static char *getopts_pos = NULL;
-
-/* ---- helpers for builtin_getopts -------------------------------------- */
-/*
- * Return the current OPTIND value as an integer.  Values less than 1 are
- * normalised to 1 and a missing variable defaults to 1 as well.
- */
-static int read_optind(void)
-{
-    const char *ind_s = get_shell_var("OPTIND");
-    int ind = ind_s ? atoi(ind_s) : 1;
-    if (ind < 1)
-        ind = 1;
-    if (!script_argv)
-        ind = 1;
-    return ind;
-}
-
-/*
- * Update the OPTIND variable maintained in the shell's variable table.
- */
-static void write_optind(int ind)
-{
-    char buf[16];
-    snprintf(buf, sizeof(buf), "%d", ind);
-    set_shell_var("OPTIND", buf);
-}
-
-/*
- * Convenience wrapper to update the OPTARG variable.  NULL is written as an
- * empty string.
- */
-static void write_optarg(const char *val)
-{
-    set_shell_var("OPTARG", val ? val : "");
-}
-
-/*
- * Helper used by getopts to place the resulting option character in the user
- * supplied variable.
- */
-static void write_result_var(const char *var, const char *val)
-{
-    set_shell_var(var, val);
-}
-
-enum {
-    OPT_OK,
-    OPT_DONE,
-    OPT_ILLEGAL,
-    OPT_MISSING
-};
-
-/*
- * Scan the argument list for the next option recognised by `getopts`.  Updates
- * OPTARG as required and returns one of the OPT_* codes defined above.
- */
-static int getopts_next_option(const char *optstr, int silent, int *ind, char *opt)
-{
-    if (!script_argv || *ind > script_argc) {
-        getopts_pos = NULL;
-        return OPT_DONE;
-    }
-
-    if (!getopts_pos || *getopts_pos == '\0') {
-        char *arg = script_argv[*ind];
-        if (strcmp(arg, "--") == 0) {
-            (*ind)++;
-            getopts_pos = NULL;
-            return OPT_DONE;
-        }
-        if (arg[0] != '-' || arg[1] == '\0') {
-            getopts_pos = NULL;
-            return OPT_DONE;
-        }
-        getopts_pos = arg + 1;
-        (*ind)++;
-    }
-
-    *opt = *getopts_pos++;
-    if (!*opt || *opt == ':')
-        *opt = '?';
-
-    const char *p = strchr(optstr, *opt);
-    if (!p) {
-        if (!silent)
-            fprintf(stderr, "getopts: illegal option -- %c\n", *opt);
-        if (silent) {
-            char ob[2] = {*opt, '\0'};
-            write_optarg(ob);
-        } else {
-            write_optarg("");
-        }
-        if (!getopts_pos || *getopts_pos == '\0')
-            getopts_pos = NULL;
-        return OPT_ILLEGAL;
-    }
-
-    if (p[1] == ':') {
-        if (*getopts_pos != '\0') {
-            write_optarg(getopts_pos);
-            getopts_pos = NULL;
-        } else if (*ind <= script_argc) {
-            write_optarg(script_argv[*ind]);
-            (*ind)++;
-            getopts_pos = NULL;
-        } else {
-            if (!silent)
-                fprintf(stderr, "getopts: option requires an argument -- %c\n", *opt);
-            if (silent) {
-                char ob[2] = {*opt, '\0'};
-                write_optarg(ob);
-            } else {
-                write_optarg("");
-            }
-            return OPT_MISSING;
-        }
-    } else {
-        write_optarg("");
-        if (!getopts_pos || *getopts_pos == '\0')
-            getopts_pos = NULL;
-    }
-    return OPT_OK;
-}
-
-/*
- * Shell builtin compatible with POSIX getopts.  Parses positional parameters
- * according to the given option string and stores the result in the supplied
- * variable.  The exit status is placed in `last_status` and the function
- * returns 1.
- */
-int builtin_getopts(char **args) {
-    if (!args[1] || !args[2]) {
-        fprintf(stderr, "usage: getopts optstring var\n");
-        last_status = 1;
-        return 1;
-    }
-
-    const char *optstr = args[1];
-    const char *var = args[2];
-    int silent = 0;
-    if (optstr[0] == ':') {
-        silent = 1;
-        optstr++;
-    }
-
-    int ind = read_optind();
-
-    char opt = '\0';
-    int res = getopts_next_option(optstr, silent, &ind, &opt);
-
-    switch (res) {
-    case OPT_OK: {
-        char val[2] = {opt, '\0'};
-        write_result_var(var, val);
-        last_status = 0;
-        break;
-    }
-    case OPT_DONE:
-        write_result_var(var, "?");
-        write_optarg("");
-        last_status = 1;
-        break;
-    case OPT_ILLEGAL:
-        write_result_var(var, "?");
-        last_status = 0;
-        break;
-    case OPT_MISSING:
-        if (silent)
-            write_result_var(var, ":");
-        else
-            write_result_var(var, "?");
-        last_status = 0;
-        break;
-    }
-
-    write_optind(ind);
-    return 1;
-}
